requirements.txt
python-dotenv~=1.0.1
flask-cors~=4.0.0
flask-sqlalchemy~=3.1.1
flask-jwt-extended~=4.6.0
psycopg2-binary~=2.9.9
Flask~=3.0.2
SQLAlchemy~=2.0.28